+++ apps/spectator/app/game/[id]/_components/CheerTalk/Banner/Banner.css.ts
