+++ apps/spectator/app/game/[id]/_components/CheerTalk/Banner/index.tsx
