+++ apps/spectator/app/game/[id]/_components/CheerTalk/EntryButton/EntryButton.css.ts
