+++ apps/spectator/app/game/[id]/_components/CheerTalk/EntryButton/index.tsx
