+++ apps/spectator/app/game/[id]/_components/CheerTalk/Item/Item.css.ts
