+++ apps/spectator/app/game/[id]/_components/CheerTalk/Item/index.tsx
