+++ apps/spectator/app/game/[id]/_components/CheerTalk/List/List.css.ts
