+++ apps/spectator/app/game/[id]/_components/CheerTalk/MenuModal/MenuModal.css.ts
