+++ apps/spectator/app/game/[id]/_components/CheerTalk/Modal/Modal.css.ts
